Fully functional lab 2 according to the test. The test code is under lab1.c and I commented out the old code of lab1.c.

+++ proc.h
@@ -51,7 +51,7 @@ struct proc {
   char name[16];               // Process name (debugging)
   int exitstatus;                   
   int priority;                 // Process priority 
-  float waitTime;
+  int waitTime;
 };
 
 // Process memory is laid out contiguously, low addresses first:
